Migra get_player_ids_from_group para o módulo player
